ci: fix image deploy (#4953)
The Influx deployment pipeline was changed so the an image push is used
as a signal for deployment (instead of a magic script that was used
before). So we need to adopt our CI to only push images when all tests
pass.

Old workflow:
- build release: builds docker images and push commit-based tags to
  registry
- deploy release: pulls built images from registry, adds+pushes branch
  tags, calls magic deploy script

New workflow:
- build release: builds docker image, saves them to disk
- deploy release: load image files, tags them, pushes tags

You may wonder why there are two steps if we could just use a single
one. The reason is: time-to-deploy. We can already build the image while we
are waiting for the tests. If the tests fail, the image will just not be
published.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>