chore: update `tracing-subscriber`, fix trogging, fix CLI test port allocation (#5581)
* test: use dedicated ports for CLI tests

* chore: update `tracing-subscriber`

* fix: work around tracing-subscriber weirdness

It seems that trogging with tracing-subscriber >= 0.3.14 does not
produce any output at all. I suspect we are hitting
<https://github.com/tokio-rs/tracing/issues/2265>. Let's change the
construct to not use multiple optional layers but a single dyn-dispatch
layer. Logging shouldn't have such a high throughput that his makes any
difference, esp. because the dyn-dispatch happens AFTER the filter.