refactor: Combine relevant data with the FilterResult state
This encodes the result directly and has the FilterResult hold only the
relevant data to the state. So no longer any need to create or check for
empty vectors or 0 budget_bytes. Also creates a new type after checking
the filter result state and handling the budget, as actual compaction
doesn't need to care about that.

This could still use more refactoring to become a clearer pipeline of
different states, but I think this is a good start.