chore: replace `ansi_style` with `nu-ansi-style` (#5669)
See <https://rustsec.org/advisories/RUSTSEC-2021-0139.html>.

This follows the migration that `tracing-subscriber` will perform:
https://github.com/tokio-rs/tracing/pull/2287

Note that `tracing-subscriber` is not released yet, so for the time
being we will have `ansi_style` and `nu-ansi-style` in our dependency
list. It's likely that `tracing-subscriber` will backport and and
release this change rather soon though, because people are getting
annoyed by RUSTSEC warnings.