refactor: parse HTTP trace header even w/o collector (#5979)
Since we log trace IDs to allow easier correlation of logs no matter
what the `sampled` flag says, we should also parse these logs if we
don't have a tace collector at all.

In practice, this won't make a difference since we always deploy with a
trace collector, but it also makes the code easier to reason about.

Helps with #5975.